feat(orb): A8.3a.2 — lift upstream message-handler body to orb/live/session/upstream-message-handler.ts (VTID-02968)
Track A — refactor / A8.3 sub-split, slice 2 of 3.

The named `handleUpstreamLiveMessage` function (created in A8.3a.1)
moves out of `routes/orb-live.ts` into a dedicated module under
`orb/live/session/`. Body is lifted verbatim except for mechanical
substitutions; behavior is byte-compatible.

A8.3 sub-split sequence:
  A8.3a.1 (shipped, #2098)  — name closure + migrate SSE writes
  A8.3a.2 (this, VTID-02968) — move named function to new module
  A8.3b   (next)            — swap connectToLiveAPI for VertexLiveClient

What this lands:
- `services/gateway/src/orb/live/session/upstream-message-handler.ts`
  (new, 1095 lines):
    * `UpstreamMessageHandlerDeps` interface — typed bag of the 10
      orb-live.ts-local helpers the body calls (clearResponseWatchdog,
      detectAuthIntent, emitDiag, emitLiveSessionEvent, executeLiveApiTool,
      isDevSandbox, sendAudioToLiveAPI, sendFunctionResponseToLiveAPI,
      sendWsMessage, startResponseWatchdog).
    * `UpstreamMessageHandlerContext` interface — captures connectToLiveAPI
      Promise-closure state (`onSetupComplete()` mutates setupComplete +
      clears connectionTimeout + resolves; `isSetupComplete()` reads the
      outer let-binding) plus the 5 user callbacks (onAudioResponse /
      onTextResponse / onError / onTurnComplete / onInterrupted).
    * `createUpstreamLiveMessageHandler(ctx)` factory — returns a
      `(data: WebSocket.Data) => void` handler bound to the context.
    * Direct module-level imports: writeSseEvent (A9.2), 7 upstream
      constants, randomUUID, emitOasisEvent, handleIdentityIntent,
      deduplicatedExtract, writeMemoryItemWithIdentity, DEV_IDENTITY,
      addSessionTurn, addTurnRedis, getSupabase, VITANA_BOT_USER_ID.
    * Body lifted verbatim from orb-live.ts:~5772-6732 with 27 mechanical
      dep substitutions (`name(` → `ctx.deps.name(`) and the 3-line
      setup-complete handshake collapsed into `ctx.onSetupComplete()`.
    * 11 SSE writes already migrated to writeSseEvent in A8.3a.1 stay
      as-is.
- `routes/orb-live.ts`:
    * 961-line inline `function handleUpstreamLiveMessage` replaced with a
      single `const handleUpstreamLiveMessage = createUpstreamLiveMessageHandler({...})`
      construction.
    * `ws.on('message', handleUpstreamLiveMessage)` registration unchanged.
    * The `onSetupComplete` wiring callback now owns the
      setupComplete-mutation / clearTimeout(connectionTimeout) /
      resolve(ws) sequence — body of the handler no longer touches them
      directly.
    * orb-live.ts shrinks ~13038 → 12108 lines.
- `test/orb/live/characterization/upstream-message-handler.characterization.test.ts`
  (updated):
    * Slice now reads from BOTH `routes/orb-live.ts` (the consumer
      seam) AND `orb/live/session/upstream-message-handler.ts` (the
      body).
    * Verifies: factory exported, inner handler named correctly,
      every event path present (setup_complete / server_content /
      tool_call / interrupted / turn_complete / transcripts), callbacks
      flow through `ctx.callbacks.*`, setup-complete handshake via
      `ctx.onSetupComplete()` (anti-regression: no inline
      `setupComplete = true` / `clearTimeout(connectionTimeout)` in the
      lifted body), `writeSseEvent` used (no inline res.write), all
      orb-live.ts-local helpers routed through `ctx.deps.*`.
    * Verifies orb-live.ts: imports the factory, builds the handler with
      the typed context, registers `ws.on('message', handleUpstreamLiveMessage)`,
      does NOT declare a `function handleUpstreamLiveMessage` body locally,
      `onSetupComplete` wiring still mutates setupComplete + clears
      connectionTimeout + resolves(ws).

Build/runtime fixes (caught + fixed during the lift):
- `WebSocket` was `import type` — needed to be value-import for
  `WebSocket.OPEN` runtime constant.
- Missing imports added: `DEV_IDENTITY`, `VITANA_BOT_USER_ID`,
  `SILENCE_IDLE_THRESHOLD_MS`, `SILENCE_KEEPALIVE_INTERVAL_MS`.
- Dynamic `import('../services/oasis-event-service')` path adjusted to
  `../../../services/oasis-event-service` for the new file location.
- Deps interface signatures relaxed to `any`-tolerant variants to absorb
  the orb-live.ts local function types (slightly tighter typing comes
  in A8.3b when the deps disappear into VertexLiveClient).
- One `[...new Set(urls)]` cast to `string[]` to satisfy strict TS.

Wire protocol: byte-for-byte identical. Same JSON envelopes through
`writeSseEvent`, same OASIS/diag payloads, same tool-call dispatch,
same transcript handling.

What this PR does NOT touch:
- `connectToLiveAPI` is still the active path (calls `ws.on('message',
  handleUpstreamLiveMessage)` against the freshly-built handler). Swap
  to `VertexLiveClient` is A8.3b.
- LiveKit adapter, L1 provider selection, F1/F4, B6/B7, R-lane.

Tests: 9 reworked structural assertions (originally 7 from A8.3a.1);
648 green across `test/orb/**` (no regression). A7 / A9.1 / A9.2 / A8.1 /
A8.2 / A8.2.1 / A8.2-complete / A8.3a.1 suites + decision-contract spine
tests + all other characterization suites still green.
